pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ae_pcc"
version = "0.1.0"
description = "Your project description here"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "numpy==1.26.4",
    "tqdm",
    "torch",
    "torchvision",
    "torchaudio",
    "torch-scatter",
    "torch-sparse",
    "torch-cluster",
    "torch-geometric",
    "trimesh",
    "open3D",
    "matplotlib",
    "tensorboard",
    "pandas",
    "click",
    "plyfile",
    "PyYAML",
    "h5py",
    "cbor2"
]

[tool.setuptools]
packages = ["architecture", "dataloaders", "loss", "utils"]